requirements.txt
asgiref==3.5.2
Django==4.1.4
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-ranged-response==0.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.2.0
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
